// give-away-vite/give-away/src/api/api.ts
export interface FashionItemDetailResponse {
    /**
     * 
     * @type {string}
     * @memberof FashionItemDetailResponse
     */
    'itemId'?: string;
    /**
     * 
     * @type {FashionItemType}
     * @memberof FashionItemDetailResponse
     */
    'type'?: FashionItemType;
    /**
     * 
     * @type {number}
     * @memberof FashionItemDetailResponse
     */
    'sellingPrice'?: number;
    /**
     * 
     * @type {string}
     * @memberof FashionItemDetailResponse
     */
    'name'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FashionItemDetailResponse
     */
    'note'?: string | null;
    /**
     * 
     * @type {number}
     * @memberof FashionItemDetailResponse
     */
    'condition'?: number;
    /**
     * 
     * @type {FashionItemStatus}
     * @memberof FashionItemDetailResponse
     */
    'status'?: FashionItemStatus;
    /**
     * 
     * @type {string}
     * @memberof FashionItemDetailResponse
     */
    'shopAddress'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FashionItemDetailResponse
     */
    'shopId'?: string;
    /**
     * 
     * @type {string}
     * @memberof FashionItemDetailResponse
     */
    'categoryName'?: string | null;
    /**
     * 
     * @type {SizeType}
     * @memberof FashionItemDetailResponse
     */
    'size'?: SizeType;
    /**
     * 
     * @type {string}
     * @memberof FashionItemDetailResponse
     */
    'color'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FashionItemDetailResponse
     */
    'brand'?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof FashionItemDetailResponse
     */
    'isOrderedYet'?: boolean;
    /**
     * 
     * @type {GenderType}
     * @memberof FashionItemDetailResponse
     */
    'gender'?: GenderType;
    /**
     * 
     * @type {Array<string>}
     * @memberof FashionItemDetailResponse
     */
    'images'?: Array<string> | null;
}
// ...
export const FashionItemStatus = {
    Available: 'Available',
    Unavailable: 'Unavailable',
    OnDelivery: 'OnDelivery',
    Sold: 'Sold',
    Refundable: 'Refundable',
    Pending: 'Pending',
    PendingAuction: 'PendingAuction',
    AwaitingAuction: 'AwaitingAuction',
    Bidding: 'Bidding',
    Won: 'Won',
    Rejected: 'Rejected',
    Returned: 'Returned'
} as const;

export type FashionItemStatus = typeof FashionItemStatus[keyof typeof FashionItemStatus];


/**
 * 
 * @export
 * @enum {string}
 */

export const FashionItemType = {
    ItemBase: 'ItemBase',
    ConsignedForSale: 'ConsignedForSale',
    ConsignedForAuction: 'ConsignedForAuction'
} as const;

export type FashionItemType = typeof FashionItemType[keyof typeof FashionItemType];


/**
 * 
 * @export
 * @enum {string}
 */

export const GenderType = {
    Male: 'Male',
    Female: 'Female'
} as const;

export type GenderType = typeof GenderType[keyof typeof GenderType];
// ...
export const SizeType = {
    Xs: 'XS',
    S: 'S',
    M: 'M',
    L: 'L',
    Xl: 'XL',
    Xxl: 'XXL',
    Xxxl: 'XXXL',
    Xxxxl: 'XXXXL'
} as const;

export type SizeType = typeof SizeType[keyof typeof SizeType];
// ...
export interface UpdateFashionItemRequest {
    /**
     * 
     * @type {number}
     * @memberof UpdateFashionItemRequest
     */
    'sellingPrice'?: number | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateFashionItemRequest
     */
    'name'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateFashionItemRequest
     */
    'note'?: string | null;
    /**
     * 
     * @type {number}
     * @memberof UpdateFashionItemRequest
     */
    'condition'?: number | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateFashionItemRequest
     */
    'brand'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateFashionItemRequest
     */
    'color'?: string | null;
    /**
     * 
     * @type {GenderType}
     * @memberof UpdateFashionItemRequest
     */
    'gender'?: GenderType;
    /**
     * 
     * @type {SizeType}
     * @memberof UpdateFashionItemRequest
     */
    'size'?: SizeType;
    /**
     * 
     * @type {string}
     * @memberof UpdateFashionItemRequest
     */
    'categoryId'?: string | null;
}
// ...
export class FashionItemApi extends BaseAPI {
    /**
     * 
     * @param {string} [searchTerm] 
     * @param {number} [pageNumber] 
     * @param {number} [pageSize] 
     * @param {string} [memberId] 
     * @param {Array<FashionItemStatus>} [status] 
     * @param {Array<FashionItemType>} [type] 
     * @param {string} [shopId] 
     * @param {GenderType} [genderType] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof FashionItemApi
     */
    public apiFashionitemsGet(searchTerm?: string, pageNumber?: number, pageSize?: number, memberId?: string, status?: Array<FashionItemStatus>, type?: Array<FashionItemType>, shopId?: string, genderType?: GenderType, options?: RawAxiosRequestConfig) {
        return FashionItemApiFp(this.configuration).apiFashionitemsGet(searchTerm, pageNumber, pageSize, memberId, status, type, shopId, genderType, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof FashionItemApi
     */
    public apiFashionitemsIdGet(id: string, options?: RawAxiosRequestConfig) {
        return FashionItemApiFp(this.configuration).apiFashionitemsIdGet(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} itemId 
     * @param {UpdateFashionItemRequest} [updateFashionItemRequest] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof FashionItemApi
     */
    public apiFashionitemsItemIdPut(itemId: string, updateFashionItemRequest?: UpdateFashionItemRequest, options?: RawAxiosRequestConfig) {
        return FashionItemApiFp(this.configuration).apiFashionitemsItemIdPut(itemId, updateFashionItemRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} itemid 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof FashionItemApi
     */
    public apiFashionitemsItemidCheckAvailabilityPut(itemid: string, options?: RawAxiosRequestConfig) {
        return FashionItemApiFp(this.configuration).apiFashionitemsItemidCheckAvailabilityPut(itemid, options).then((request) => request(this.axios, this.basePath));
    }
}